Drop preferences/type.ts, move all types to preferences/db
There was a duplication of types between preferences/types and
preferences/db, mostly due to a missed duplication way back when the
preferences service was still being booted up. This commit eliminates
the duplication, harmonizes the types, and re-exports appropriate
external-facing types.

No underlying functionality has changed.
