+++ background/redux-slices/ui.ts
@@ -4,8 +4,7 @@ import { AddressOnNetwork } from "../accounts"
 import { ETHEREUM } from "../constants"
 import { AnalyticsEvent, OneTimeAnalyticsEvent } from "../lib/posthog"
 import { EVMNetwork } from "../networks"
-import { DismissableItem } from "../services/preferences"
-import { AnalyticsPreferences } from "../services/preferences/types"
+import { AnalyticsPreferences, DismissableItem } from "../services/preferences"
 import { AccountSignerWithId } from "../signing"
 import { AccountSignerSettings } from "../ui"
 import { AccountState, addAddressNetwork } from "./accounts"
