+++ background/services/preferences/defaults.ts
@@ -1,10 +1,9 @@
 import { ETHEREUM, MINUTE, USD } from "../../constants"
 import { storageGatewayURL } from "../../lib/storage-gateway"
-import { Preferences } from "./types"
 
 export const DEFAULT_AUTOLOCK_INTERVAL = 60 * MINUTE
 
-const defaultPreferences: Preferences = {
+const defaultPreferences = {
   tokenLists: {
     autoUpdate: false,
     urls: [
